+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Fixed
 
+- Fix `displayTabContentForItem` for `PluginDatabaseinventoryContactLog`
 - Fix foreign key constraint in `where` clause.
 
 ## [1.0.1] - 2024-12-11